znc_search/requirements.txt
blinker==1.9.0
click==8.3.1
Flask==3.0.0
Flask-Cors==4.0.0
gunicorn==21.2.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==25.0
pillow==12.1.0
pyotp==2.9.0
pysqlcipher3==1.2.0
python-dotenv==1.2.1
qrcode==8.2
Werkzeug==3.1.5
